refactor: unify session ID generation and fix tool name length violations
- Centralize session ID generation in `shared/utils.ts` using `nanoid`
  (12 characters, starts with a letter per OpenAI requirements)
- Update all storage backends (File, Memory, Redis, SQLite, Supabase) to
  use the shared `generateSessionId` utility to remove code duplication
- Fix "string too long" error for tool names by truncating the
  `serverId`-derived prefix to 8 characters across all SDK adapters
  (AI, AG-UI, LangChain, Mastra)
- Export `generateSessionId` from `server/storage/index.ts`
