+++ src/adapters/langchain-adapter.ts
@@ -58,7 +58,7 @@ export class LangChainAdapter {
         await this.ensureDependencies();
 
         const result = await client.listTools();
-        const prefix = this.options.prefix ?? client.getServerId() ?? 'mcp';
+        const prefix = this.options.prefix ?? client.getServerId()?.replace(/-/g, '').substring(0, 8) ?? 'mcp';
 
         return result.tools.map((tool) => {
             // In a real implementation, you would use a library like 'json-schema-to-zod'
